# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}
export type AnalyzeMaterialInputKind = typeof AnalyzeMaterialInputKind[keyof typeof AnalyzeMaterialInputKind];


export const AnalyzeMaterialInputKind = {
  syllabus: 'syllabus',
  lecture: 'lecture',
} as const;

export interface AnalyzeMaterialInput {
  kind: AnalyzeMaterialInputKind;
  filename: string;
  courseName?: string;
  /** Base64-encoded PDF content. Never store this value. */
  contentBase64?: string;
  /** Optional extracted text for clients that already have a text layer. */
  text?: string;
}

export type AnalyzeMaterialOutputKind = typeof AnalyzeMaterialOutputKind[keyof typeof AnalyzeMaterialOutputKind];


export const AnalyzeMaterialOutputKind = {
  syllabus: 'syllabus',
  lecture: 'lecture',
} as const;

export type AnalyzeMaterialOutputCourseSuggestion = {
  name: string;
  code: string;
  instructor?: string;
};

export interface AnalyzedTopic {
  title: string;
  description: string;
  confidence: number;
  evidence: string;
}

export interface AnalyzedDeadline {
  title: string;
  type: string;
  dueDate: string;
  confidence: number;
  evidence: string;
}

export interface StudySuggestion {
  title: string;
  reason: string;
  estimatedMinutes: number;
}

export interface AnalyzeMaterialOutput {
  kind: AnalyzeMaterialOutputKind;
  filename: string;
  summary: string;
  courseSuggestion?: AnalyzeMaterialOutputCourseSuggestion;
  topics: AnalyzedTopic[];
  deadlines: AnalyzedDeadline[];
  studySuggestions: StudySuggestion[];
  warnings: string[];
  sourceTextChars: number;
}