# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ai
    description: Secure AI-powered academic material analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /ai/analyze-material:
    post:
      operationId: analyzeAcademicMaterial
      tags: [ai]
      summary: Analyze a syllabus or lecture PDF
      description: Extracts reviewable academic findings through the server-side AI service. Findings are suggestions and require user confirmation before becoming course data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeMaterialInput"
      responses:
        "200":
          description: Structured findings for user review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeMaterialOutput"
        "400":
          description: Invalid material or unsupported file
        "502":
          description: AI analysis failed
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnalyzeMaterialInput:
      type: object
      properties:
        kind:
          type: string
          enum: [syllabus, lecture]
        filename:
          type: string
        courseName:
          type: string
        contentBase64:
          type: string
          description: Base64-encoded PDF content. Never store this value.
        text:
          type: string
          description: Optional extracted text for clients that already have a text layer.
      required:
        - kind
        - filename
    AnalyzeMaterialOutput:
      type: object
      properties:
        kind:
          type: string
          enum: [syllabus, lecture]
        filename:
          type: string
        summary:
          type: string
        courseSuggestion:
          type: object
          properties:
            name: { type: string }
            code: { type: string }
            instructor: { type: string }
          required: [name, code]
        topics:
          type: array
          items:
            $ref: "#/components/schemas/AnalyzedTopic"
        deadlines:
          type: array
          items:
            $ref: "#/components/schemas/AnalyzedDeadline"
        studySuggestions:
          type: array
          items:
            $ref: "#/components/schemas/StudySuggestion"
        warnings:
          type: array
          items: { type: string }
        sourceTextChars:
          type: integer
      required: [kind, filename, summary, topics, deadlines, studySuggestions, warnings, sourceTextChars]
    AnalyzedTopic:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        confidence: { type: number }
        evidence: { type: string }
      required: [title, description, confidence, evidence]
    AnalyzedDeadline:
      type: object
      properties:
        title: { type: string }
        type: { type: string }
        dueDate: { type: string }
        confidence: { type: number }
        evidence: { type: string }
      required: [title, type, dueDate, confidence, evidence]
    StudySuggestion:
      type: object
      properties:
        title: { type: string }
        reason: { type: string }
        estimatedMinutes: { type: integer }
      required: [title, reason, estimatedMinutes]