# lib/api-zod/src/generated/types/analyzeMaterialInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeMaterialInputKind } from './analyzeMaterialInputKind';

export interface AnalyzeMaterialInput {
  kind: AnalyzeMaterialInputKind;
  filename: string;
  courseName?: string;
  /** Base64-encoded PDF content. Never store this value. */
  contentBase64?: string;
  /** Optional extracted text for clients that already have a text layer. */
  text?: string;
}