# lib/api-zod/src/generated/types/analyzeMaterialOutput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzedDeadline } from './analyzedDeadline';
import type { AnalyzedTopic } from './analyzedTopic';
import type { AnalyzeMaterialOutputCourseSuggestion } from './analyzeMaterialOutputCourseSuggestion';
import type { AnalyzeMaterialOutputKind } from './analyzeMaterialOutputKind';
import type { StudySuggestion } from './studySuggestion';

export interface AnalyzeMaterialOutput {
  kind: AnalyzeMaterialOutputKind;
  filename: string;
  summary: string;
  courseSuggestion?: AnalyzeMaterialOutputCourseSuggestion;
  topics: AnalyzedTopic[];
  deadlines: AnalyzedDeadline[];
  studySuggestions: StudySuggestion[];
  warnings: string[];
  sourceTextChars: number;
}