# lib/api-zod/src/generated/types/analyzedDeadline.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzedDeadline {
  title: string;
  type: string;
  dueDate: string;
  confidence: number;
  evidence: string;
}