# lib/api-zod/src/generated/types/analyzedTopic.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzedTopic {
  title: string;
  description: string;
  confidence: number;
  evidence: string;
}