# lib/api-zod/src/generated/types/studySuggestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudySuggestion {
  title: string;
  reason: string;
  estimatedMinutes: number;
}